user class moved to model folder
